+++ crayg/intersectors/SceneIntersector.h
@@ -9,13 +9,22 @@
 #include <scene/Scene.h>
 #include "Intersectable.h"
 
+struct Intersection{
+public:
+    Intersection(const Ray &ray, Vector3f location, SceneObject *object) : ray(ray), location(location), object(object) {}
+
+    Ray ray;
+    Vector3f location;
+    SceneObject* object;
+};
+
 // todo add init method -> builds BVH Tree
-class SceneIntersector : public Intersectable{
+class SceneIntersector{
 public:
     explicit SceneIntersector(Scene &scene);
 
-    Vector3f intersect(Ray ray) override;
-    bool intersects(Ray ray) override;
+    Intersection intersect(Ray ray);
+    bool intersects(Ray ray);
 
 private:
     Scene &scene;